requirements.txt
accelerate>=0.25.0
diffusers==0.33.1
draccus==0.8.0
einops
einx
huggingface_hub
json-numpy
jsonlines
matplotlib
peft==0.11.1
protobuf
rich
sentencepiece==0.1.99
timm==0.9.10
tokenizers==0.19.1
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
transformers==4.40.1
wandb
tensorflow==2.15.0
tensorflow_datasets==4.9.3
tensorflow_graphics==2021.12.3
dlimp @ git+https://github.com/moojink/dlimp_openvla
black>=24.2.0
gpustat
ipython
pre-commit
ruff>=0.2.2
boto3
sagemaker
positional-encodings==6.0.3
hydra-core==1.1.1
tensorboardX